pyproject.toml
[tool.poetry]
name = "lectionary-pdf-generator"
version = "0.1.0"
description = ""
authors = ["tobycheng <tobycfcheng@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "^2.146.0"
constructs = "^10.3.0"
requests = "^2.32.3"
bs4 = "^0.0.2"
pymupdf = "^1.24.5"
boto3 = "^1.34.130"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
